lab_3/user_service/requirements.txt
grpcio ~= 1.50
grpcio-tools ~= 1.50
sqlalchemy ~= 2.0

